split tests into groups
