+++ .github/workflows/continuous-integration-workflow.yml
@@ -14,20 +14,20 @@ jobs:
       - uses: actions/setup-node@v3
         with:
           node-version: 20
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Run code formatting check
         run: npm run fmt:check
-  playwright-tests:
-    name: Playwright tests
+  playwright-tests-blog:
+    name: Blog - Playwright tests
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
         with:
           node-version: 20
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: |
           npm ci
@@ -36,5 +36,76 @@ jobs:
           npx playwright install
       - name: Run tests
         run: |
-          npx playwright test
-
+          npx playwright test playwright-tests/tests/blog
+  playwright-tests-community:
+    name: Community - Playwright tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 20
+          cache: "npm"
+      - name: Install dependencies
+        run: |
+          npm ci
+          curl --proto '=https' --tlsv1.2 -LsSf https://github.com/mpeterdev/bos-loader/releases/download/v0.7.1/bos-loader-v0.7.1-installer.sh | sh
+          npx playwright install-deps
+          npx playwright install
+      - name: Run tests
+        run: |
+          npx playwright test playwright-tests/tests/community
+  playwright-tests-other:
+    name: Other - Playwright tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 20
+          cache: "npm"
+      - name: Install dependencies
+        run: |
+          npm ci
+          curl --proto '=https' --tlsv1.2 -LsSf https://github.com/mpeterdev/bos-loader/releases/download/v0.7.1/bos-loader-v0.7.1-installer.sh | sh
+          npx playwright install-deps
+          npx playwright install
+      - name: Run tests
+        run: |
+          npx playwright test playwright-tests/tests/other
+  playwright-tests-proposal:
+    name: Proposal - Playwright tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 20
+          cache: "npm"
+      - name: Install dependencies
+        run: |
+          npm ci
+          curl --proto '=https' --tlsv1.2 -LsSf https://github.com/mpeterdev/bos-loader/releases/download/v0.7.1/bos-loader-v0.7.1-installer.sh | sh
+          npx playwright install-deps
+          npx playwright install
+      - name: Run tests
+        run: |
+          npx playwright test playwright-tests/tests/proposal
+  playwright-tests-sunset:
+    name: Sunset - Playwright tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 20
+          cache: "npm"
+      - name: Install dependencies
+        run: |
+          npm ci
+          curl --proto '=https' --tlsv1.2 -LsSf https://github.com/mpeterdev/bos-loader/releases/download/v0.7.1/bos-loader-v0.7.1-installer.sh | sh
+          npx playwright install-deps
+          npx playwright install
+      - name: Run tests
+        run: |
+          npx playwright test playwright-tests/tests/sunset
