+++ playwright-tests/tests/other/bosloaderenvironment.spec.js
