+++ playwright-tests/tests/sunset/feed.spec.js
@@ -1,5 +1,5 @@
 import { expect, test } from "@playwright/test";
-import { pauseIfVideoRecording } from "../testUtils.js";
+import { pauseIfVideoRecording } from "../../testUtils.js";
 
 test("should show post history for posts in the feed", async ({ page }) => {
   await page.goto("/devhub.near/widget/app?page=feed");
