+++ playwright-tests/tests/sunset/funding.spec.js
@@ -3,7 +3,7 @@ import {
   pauseIfVideoRecording,
   selectAndAssert,
   setInputAndAssert,
-} from "../testUtils";
+} from "../../testUtils";
 
 test.describe("Wallet is connected", () => {
   // sign in to wallet
